requirements/ci.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
asgiref==3.4.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
astroid==2.8.0
    # via pylint
beautifulsoup4==4.10.0
    # via webtest
certifi==2021.5.30
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   elastic-apm
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.6
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
coverage==4.5.4
    # via -r requirements/test-tools.in
cssselect==1.1.0
    # via pyquery
django==3.2.7
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
    #   django-choices
    #   django-formtools
    #   django-hijack
    #   django-localflavor
    #   django-otp
    #   django-phonenumber-field
    #   django-privates
    #   django-redis
    #   django-rosetta
    #   django-sendfile2
    #   django-sniplates
    #   maykin-django-two-factor-auth
django-admin-index==1.5.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-axes==5.25.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-choices==1.7.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-formtools==2.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
django-hijack==3.0.4
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-ipware==4.0.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
django-localflavor==3.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-ordered-model==3.4.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-admin-index
django-otp==1.1.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
django-phonenumber-field==5.2.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
django-privates==1.2.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-redis==5.0.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-rosetta==0.9.7
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-sendfile2==0.6.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-privates
django-sniplates==0.7.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-webtest==1.9.7
    # via -r requirements/test-tools.in
elastic-apm==6.4.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
factory-boy==3.2.0
    # via -r requirements/test-tools.in
faker==8.14.1
    # via factory-boy
freezegun==1.1.0
    # via -r requirements/test-tools.in
idna==3.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
isort==5.9.3
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
lxml==4.6.3
    # via pyquery
maykin-django-two-factor-auth==2.0.4
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
mccabe==0.6.1
    # via pylint
pep8==1.7.1
    # via -r requirements/test-tools.in
phonenumbers==8.12.33
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
pillow==8.3.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
platformdirs==2.4.0
    # via pylint
polib==1.1.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-rosetta
psycopg2==2.9.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
pylint==2.11.1
    # via -r requirements/test-tools.in
pyquery==1.4.3
    # via -r requirements/test-tools.in
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
python-decouple==3.5
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
python-dotenv==0.19.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
python-stdnum==1.17
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-localflavor
pytz==2021.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
qrcode==6.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
redis==3.5.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-redis
requests==2.26.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-rosetta
    #   requests-mock
requests-mock==1.9.3
    # via -r requirements/test-tools.in
sentry-sdk==1.4.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
six==1.16.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-choices
    #   python-dateutil
    #   qrcode
    #   requests-mock
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
tblib==1.7.0
    # via -r requirements/test-tools.in
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pylint
typing-extensions==3.10.0.2
    # via
    #   astroid
    #   pylint
urllib3==1.26.7
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
waitress==2.0.0
    # via webtest
webob==1.8.7
    # via webtest
webtest==3.0.0
    # via django-webtest
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
